Feature/sprint 4 newupdates

+++ Project/Project/Interface/IMainPage.cs
@@ -22,7 +22,7 @@ namespace Project.Interface
 
         int minuteRange { get; set; }
 
-        int minuteNotifyEvery { get; set; }
+        double minuteNotifyEvery { get; set; }
 
         NotifyIcon notifyIcon { get; set; }
